de todo
gt

+++ database/migrations/2019_02_13_141221_create_repuestos_table.php
@@ -87,8 +87,8 @@ class CreateRepuestosTable extends Migration
 
         Schema::create('precios_historicos', function (Blueprint $table) {
             $table->increments('id');
-            $table->float('precio_id');
-            $table->float('repuesto_id');
+            $table->unsignedInteger('precio_id');
+            $table->unsignedInteger('repuesto_id');
 
             $table->foreign('precio_id')
                 ->references('id')
